Input order of a task is now sorted according to the order of definition of the connections (#251)
* test config graph nodes and edges id
this was done to make sure the edge ids in the graph depend on the order the cnx are added in the ron config
since this is true, this can be used to sort the input order according to the cnx order

* task input order is now related to the connection order

Previously, the input order of a tasks was sorted according to the order
in which the input nodes are created in the runtime plan(which means
that it was related to the order the nodes were defined in the config
file).
Now, the input order of a task is related to the order of the
connections in the cnx array in the config file.
The inputs are actually sorted according to the edge id in the config
graph, that is assigned whenever an edge is created. Since the edges are
created in the same order as the cnx array(or when a connection is
created programatically), sorting by the edge id achieves the desired
behaviour.

* typo