Adding a basic ELRS/CRSF bridge. (#498)
* Adding a basic ELRS/CRSF bridge.

This is again useful in the drone space to send telemetry or remote
control over the waves at long distance (up to the curvature of the
earth in line of sight)

* snafu in pure copper logging.

* Moved route on cubridges to optional

lots of bridges do have a natural ID -> Route map and suddenly this
mandatory route parameted make them stutturing

for example an AIO ESC have 4 motors, you cannot route anything but the
API was forcing you to add a random string to it.

This will be followed by some helper to generate more complex bridges
like middleware ones where the user needs to generate it at the last
minute with all the types they want to expose to each ID.

* Bunch of quality of life improvements for CuBridges

* merge snafu

* missing a spot on baremetal

* Fix for id case handling mismatched between the 2 macros

* Moved the route parameter on CuBridge to optional (#499)

* Moved route on cubridges to optional

lots of bridges do have a natural ID -> Route map and suddenly this
mandatory route parameted make them stutturing

for example an AIO ESC have 4 motors, you cannot route anything but the
API was forcing you to add a random string to it.

This will be followed by some helper to generate more complex bridges
like middleware ones where the user needs to generate it at the last
minute with all the types they want to expose to each ID.

* removing some AI stupidity

* Mopping up dumb errors from CI.

* Moved route on cubridges to optional

lots of bridges do have a natural ID -> Route map and suddenly this
mandatory route parameted make them stutturing

for example an AIO ESC have 4 motors, you cannot route anything but the
API was forcing you to add a random string to it.

This will be followed by some helper to generate more complex bridges
like middleware ones where the user needs to generate it at the last
minute with all the types they want to expose to each ID.

* Bunch of quality of life improvements for CuBridges

* merge snafu

* missing a spot on baremetal

* Fix for id case handling mismatched between the 2 macros

* WIP

* fixes with new API

* pass integrating to the bdshot demo

* Should be a good base for an end to end demo with elrs

* renaming of crates

* fix dep versions

* missed a CI spot for embedded

* dep cleanup